# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Interview Guide Platform API",
    "description": "High-performance Rust Axum API for Interview Guide & Technical Knowledge Platform",
    "contact": {
      "name": "Platform Engineering",
      "email": "eng@interviewguide.internal"
    },
    "license": {
      "name": ""
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Exchange the HTTP-only refresh cookie for a new access token (rotates the refresh token).",
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "New access token; refresh cookie rotated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.error.ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/webauthn/login/finish": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "POST /api/auth/webauthn/login/finish",
        "operationId": "login_finish",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyLoginFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Passkey authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/api/auth/webauthn/login/start": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "POST /api/auth/webauthn/login/start",
        "operationId": "login_start",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyLoginStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Passkey login challenge generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyLoginStartResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/webauthn/register/finish": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "POST /api/auth/webauthn/register/finish",
        "operationId": "register_finish",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyRegisterFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Passkey credential registered successfully"
          },
          "400": {
            "description": "Verification failed"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/auth/webauthn/register/start": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "POST /api/auth/webauthn/register/start",
        "operationId": "register_start",
        "responses": {
          "200": {
            "description": "Passkey registration challenge generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyRegisterStartResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "user"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "expires_in": {
            "type": "integer",
            "format": "int64",
            "description": "Access token lifetime in seconds. Refresh via `POST /api/auth/refresh` (HTTP-only cookie)."
          },
          "token_type": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserProfile"
          }
        }
      },
...
      "PasskeyLoginFinishRequest": {
        "type": "object",
        "required": [
          "challenge_id",
          "credential"
        ],
        "properties": {
          "challenge_id": {
            "type": "string"
          },
          "credential": {}
        }
      },
      "PasskeyLoginStartRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PasskeyLoginStartResponse": {
        "type": "object",
        "required": [
          "challenge_id",
          "public_key_credential_request_options"
        ],
        "properties": {
          "challenge_id": {
            "type": "string"
          },
          "public_key_credential_request_options": {}
        }
      },
      "PasskeyRegisterFinishRequest": {
        "type": "object",
        "required": [
          "challenge_id",
          "credential"
        ],
        "properties": {
          "challenge_id": {
            "type": "string"
          },
          "credential": {},
          "nickname": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PasskeyRegisterStartResponse": {
        "type": "object",
        "required": [
          "challenge_id",
          "public_key_credential_creation_options"
        ],
        "properties": {
          "challenge_id": {
            "type": "string"
          },
          "public_key_credential_creation_options": {}
        }
      },
...
      "UserProfile": {
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string"
          },
          "full_name": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}